Check the versions of npm and node before each installation
